Add tests for Yaml definition loaders
